requirements.txt
# FastAPI and Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Redis and Caching
redis==5.0.1
hiredis==2.2.3

# Celery for Async Tasks
celery==5.3.4
flower==2.0.1

# OCR and PDF Processing
paddleocr==2.8.1
paddlepaddle==3.0.0
pdf2image==1.17.0
Pillow==10.4.0

# AI Integration
openai==1.55.3
anthropic==0.39.0

# Data Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Monitoring and Logging
structlog==24.4.0
sentry-sdk[fastapi]==2.18.0
prometheus-client==0.21.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1

# Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0
